mo_prestamo/requirements.txt
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-environ==0.10.0
mysqlclient==2.2.0
PyMySQL==1.1.0